feat: add settings screen
- Added appearance settings
- Added update check functionality
- Added about button
- Improved settings screen UI
- Created reusable components
